# docs/swagger.yaml
basePath: /
definitions:
  github_com_opsmx_ai-guardian-api_pkg_client.CreateIntegrationRequest:
    properties:
      category:
        type: string
      featureConfigs:
        additionalProperties: true
        type: object
      id:
        type: string
      integratorConfigs:
        additionalProperties: true
        type: object
      integratorType:
        type: string
      name:
        type: string
      team:
        items:
          $ref: '#/definitions/github_com_opsmx_ai-guardian-api_pkg_client.TeamAssignment'
        type: array
    type: object
  github_com_opsmx_ai-guardian-api_pkg_client.TeamAssignment:
    properties:
      teamId:
        type: string
      teamName:
        type: string
    type: object
...
  github_com_opsmx_ai-guardian-api_pkg_service.CreateGitHubIntegrationRequest:
    properties:
      installationId:
        type: string
      name:
        type: string
      team_ids:
        items:
          type: string
        type: array
      timestamp:
        description: for app based access token decryption
        type: integer
      token:
        type: string
    type: object
...
  github_com_opsmx_ai-guardian-api_pkg_service.UpdateGitHubIntegrationRequest:
    properties:
      id:
        type: string
      installationId:
        type: string
      name:
        type: string
      team_ids:
        items:
          type: string
        type: array
      timestamp:
        description: for app based access token decryption
        type: integer
      token:
        type: string
    type: object
...
  github_com_opsmx_ai-guardian-api_pkg_service.ValidateGitHubIntegrationRequest:
    properties:
      name:
        type: string
      team_ids:
        items:
          type: string
        type: array
      token:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: AI Guardian API Service
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: AI Guardian API
  version: "1.0"
paths:
...
  /api/v1/integrations:
    get:
      consumes:
      - application/json
      description: Returns a list of all active integrations
      parameters:
      - description: Comma-separated list of team IDs to filter by
        in: query
        name: teamIds
        type: string
      - description: Integrator type filter
        in: query
        name: integratorType
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: List active integrations
      tags:
      - Integrations
  /api/v1/integrations/cloud:
    post:
      consumes:
      - application/json
      description: Creates a new cloud integration for the given `teamId`.
      parameters:
      - description: Team ID
        in: query
        name: teamId
        required: true
        type: string
      - description: Cloud integration request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_opsmx_ai-guardian-api_pkg_client.CreateIntegrationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body or missing required parameters
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create cloud integration
      tags:
      - Integrations
    put:
      consumes:
      - application/json
      description: Updates a cloud integration for the given `teamId`. `integratorType`
        must be present in the request body.
      parameters:
      - description: Team ID
        in: query
        name: teamId
        required: true
        type: string
      - description: Cloud integration request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_opsmx_ai-guardian-api_pkg_client.CreateIntegrationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body or missing required parameters
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update cloud integration
      tags:
      - Integrations
...
  /api/v1/integrations/github/create:
    post:
      consumes:
      - application/json
      description: Creates a new GitHub integration with the provided details
      parameters:
      - description: GitHub integration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_opsmx_ai-guardian-api_pkg_service.CreateGitHubIntegrationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create GitHub integration
      tags:
      - Integrations
  /api/v1/integrations/github/delete:
    delete:
      consumes:
      - application/json
      description: Deletes an integration by `integrationId` and `integrationName`
        within the provided `hubID`.
      parameters:
      - description: Integration ID
        in: query
        name: integrationId
        required: true
        type: string
      - description: Integration name
        in: query
        name: integrationName
        required: true
        type: string
      - description: Hub ID
        in: query
        name: hubID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Missing required parameters
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete integration
      tags:
      - Integrations
  /api/v1/integrations/github/details:
    get:
      consumes:
      - '*/*'
      description: Github Integrations details via Github APIs
      parameters:
      - description: Account ID (e.g., 0x3b32)
        in: query
        name: accountId
        required: true
        type: string
      - description: Integration name
        in: query
        name: integrationName
        required: true
        type: string
      - description: Hub ID
        in: query
        name: hubId
        required: true
        type: string
      - description: Repository name (optional)
        in: query
        name: repository
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Github Integrations details via Github APIs
      tags:
      - Integrations
  /api/v1/integrations/github/install:
    get:
      consumes:
      - '*/*'
      description: Returns the GitHub App installation URL.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Install GitHub App
      tags:
      - Integrations
...
  /api/v1/integrations/github/update:
    put:
      consumes:
      - application/json
      description: Updates a GitHub integration with the provided details
      parameters:
      - description: GitHub integration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_opsmx_ai-guardian-api_pkg_service.UpdateGitHubIntegrationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update GitHub integration
      tags:
      - Integrations
  /api/v1/integrations/github/validate:
    post:
      consumes:
      - application/json
      description: Validates the provided GitHub integration credentials
      parameters:
      - description: GitHub integration credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_opsmx_ai-guardian-api_pkg_service.ValidateGitHubIntegrationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Validate GitHub integration
      tags:
      - Integrations
  /api/v1/integrations/team:
    get:
      consumes:
      - application/json
      description: Returns a list of all integrations for a team
      parameters:
      - description: Team ID
        in: query
        name: teamId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: List integrations for a team
      tags:
      - Integrations
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"