# pkg/controller/integrator/integrator.go
package integrator

import (
	"encoding/json"
	"net/http"
	"strings"

	"github.com/opsmx/ai-guardian-api/pkg/client"
	"github.com/opsmx/ai-guardian-api/pkg/service"
	"github.com/opsmx/ai-guardian-api/pkg/utils"
)

type IntegratorController struct {
	integratorService *service.IntegrationService
	logger            *utils.ErrorLogger
}

func NewIntegratorController() *IntegratorController {
	return &IntegratorController{
		integratorService: service.NewIntegrationService(),
		logger:            utils.NewErrorLogger("integrator_controller"),
	}
}

// CreateGitHubIntegration creates a new GitHub integration
// @Summary Create GitHub integration
// @Description Creates a new GitHub integration with the provided details
// @Tags Integrations
// @Accept  json
// @Produce  json
// @Param   request body service.CreateGitHubIntegrationRequest true "GitHub integration details"
// @Success 201 {object} map[string]interface{}
// @Failure 400 {object} map[string]string "Invalid request body"
// @Failure 500 {object} map[string]string "Internal server error"
// @Security ApiKeyAuth
// @Router /api/v1/integrations/github/create [post]
func (c *IntegratorController) CreateGitHubIntegration(w http.ResponseWriter, r *http.Request) {
	var req service.CreateGitHubIntegrationRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		c.logger.LogError(err, "Failed to decode request", nil)
		http.Error(w, "Invalid request body", http.StatusBadRequest)
		return
	}

	integration, err := c.integratorService.CreateGitHubIntegration(r.Context(), req)
	if err != nil {
		c.logger.LogError(err, "Failed to create integration", nil)
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}

	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusCreated)
	json.NewEncoder(w).Encode(map[string]interface{}{
		"success": true,
		"data":    integration,
	})
}

// UpdateGitHubIntegration updates a GitHub integration
// @Summary Update GitHub integration
// @Description Updates a GitHub integration with the provided details
// @Tags Integrations
// @Accept  json
// @Produce  json
// @Param   request body service.UpdateGitHubIntegrationRequest true "GitHub integration details"
// @Success 200 {object} map[string]interface{}
// @Failure 400 {object} map[string]string "Invalid request body"
// @Failure 500 {object} map[string]string "Internal server error"
// @Security ApiKeyAuth
// @Router /api/v1/integrations/github/update [put]
func (c *IntegratorController) UpdateGitHubIntegration(w http.ResponseWriter, r *http.Request) {
	var req service.UpdateGitHubIntegrationRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		utils.SendErrorResponse(w, http.StatusBadRequest, "Invalid request body")
		return
	}

	integration, err := c.integratorService.UpdateGitHubIntegration(r.Context(), req)
	if err != nil {
		c.logger.LogError(err, "Failed to update GitHub integration", nil)
		utils.SendErrorResponse(w, http.StatusInternalServerError, err.Error())
		return
	}

	utils.SendSuccessResponse(w, integration, "GitHub integration updated successfully")
}

// ValidateGitHubIntegration validates GitHub integration credentials
// @Summary Validate GitHub integration
// @Description Validates the provided GitHub integration credentials
// @Tags Integrations
// @Accept  json
// @Produce  json
// @Param   request body service.ValidateGitHubIntegrationRequest true "GitHub integration credentials"
// @Success 200 {object} map[string]interface{}
// @Failure 400 {object} map[string]string "Invalid request body"
// @Failure 401 {object} map[string]string "Unauthorized"
// @Failure 500 {object} map[string]string "Internal server error"
// @Security ApiKeyAuth
// @Router /api/v1/integrations/github/validate [post]
func (c *IntegratorController) ValidateGitHubIntegration(w http.ResponseWriter, r *http.Request) {
	var req service.ValidateGitHubIntegrationRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		c.logger.LogError(err, "Failed to decode request", nil)
		http.Error(w, "Invalid request body", http.StatusBadRequest)
		return
	}
	validation, err := c.integratorService.ValidateGitHubIntegration(r.Context(), req)
	if err != nil {
		c.logger.LogError(err, "Failed to validate GitHub integration", nil)
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)
	json.NewEncoder(w).Encode(map[string]interface{}{
		"success": true,
		"data":    validation,
	})
}

// InstallGitHubAppIntegration installs the GitHub App integration
// @Summary Install GitHub App
// @Description Returns the GitHub App installation URL.
// @Tags Integrations
// @Accept */*
// @Produce  json
// @Success 200 {object} map[string]interface{}
// @Failure 500 {object} map[string]string "Internal server error"
// @Security ApiKeyAuth
// @Router /api/v1/integrations/github/install [get]
func (c *IntegratorController) InstallGitHubAppIntegration(w http.ResponseWriter, r *http.Request) {

	installationUrl, err := c.integratorService.GetGithubAppInstallationURL(r.Context())
	if err != nil {
		c.logger.LogError(err, "Failed to get GitHub app install url", nil)
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)
	redirectionURL := struct {
		Url string `json:"url"`
	}{
		Url: installationUrl,
	}
	json.NewEncoder(w).Encode(map[string]interface{}{
		"success": true,
		"data":    redirectionURL,
	})
}

// GetIntegrationsGithubDetails get GitHub integration details based on various params
// @Summary Github Integrations details via Github APIs
// @Description Github Integrations details via Github APIs
// @Tags Integrations
// @Accept */*
// @Produce  json
// @Param accountId query string true "Account ID (e.g., 0x3b32)"
// @Param integrationName query string true "Integration name"
// @Param hubId query string true "Hub ID"
// @Param repository query string false "Repository name (optional)"
// @Success 200 {object} map[string]interface{}
// @Failure 400 {object} map[string]string "Invalid request"
// @Failure 401 {object} map[string]string "Unauthorized"
// @Failure 500 {object} map[string]string "Internal server error"
// @Security ApiKeyAuth
// @Router /api/v1/integrations/github/details [get]
func (c *IntegratorController) GetIntegrationsGithubDetails(w http.ResponseWriter, r *http.Request) {
	integrationId := r.URL.Query().Get("accountId")
	if integrationId == "" {
		utils.SendErrorResponse(w, http.StatusBadRequest, "Account ID is required")
		return
	}

	integrationName := r.URL.Query().Get("integrationName")
	if integrationName == "" {
		utils.SendErrorResponse(w, http.StatusBadRequest, "Integration name is required")
		return
	}

	hubID := r.URL.Query().Get("hubId")
	if hubID == "" {
		utils.SendErrorResponse(w, http.StatusBadRequest, "Hub ID is required")
		return
	}

	repository := r.URL.Query().Get("repository")
	queryParams := map[string]string{
		// automated param from UI
		"accountId": integrationId,

		// default params
		// ssd will look for repos from installation id based token
		// if orgName is blank
		"platform":  "github", // automate platform in future release
		"scanLevel": "repository",
		"type":      "user",
	}

	details, err := c.integratorService.GetGithubIntegrationsDetails(r.Context(), queryParams, integrationId, integrationName, hubID, repository)
	if err != nil {
		c.logger.LogError(err, "Failed to get github integrations details", nil)
		utils.SendErrorResponse(w, http.StatusNotFound, err.Error())
		return
	}

	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)
	json.NewEncoder(w).Encode(map[string]interface{}{
		"success": true,
		"data":    details,
	})
}

// CreateCloudIntegration creates a new cloud integration for a team.
// @Summary Create cloud integration
// @Description Creates a new cloud integration for the given `teamId`.
// @Tags Integrations
// @Accept json
// @Produce json
// @Param teamId query string true "Team ID"
// @Param request body client.CreateIntegrationRequest true "Cloud integration request"
// @Success 200 {object} map[string]interface{}
// @Failure 400 {object} map[string]interface{} "Invalid request body or missing required parameters"
// @Failure 500 {object} map[string]interface{} "Internal server error"
// @Security ApiKeyAuth
// @Router /api/v1/integrations/cloud [post]
func (c *IntegratorController) CreateCloudIntegration(w http.ResponseWriter, r *http.Request) {
	teamID := r.URL.Query().Get("teamId")
	if teamID == "" {
		utils.SendErrorResponse(w, http.StatusBadRequest, "teamId is required")
		return
	}

	var req client.CreateIntegrationRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		c.logger.LogError(err, "failed to decode cloud integration request", nil)
		utils.SendErrorResponse(w, http.StatusBadRequest, "invalid request body")
		return
	}

	msg, err := c.integratorService.CreateCloudIntegration(r.Context(), teamID, &req)
	if err != nil {
		c.logger.LogError(err, "failed to create cloud integration", nil)
		utils.SendErrorResponse(w, http.StatusInternalServerError, err.Error())
		return
	}

	utils.SendSuccessResponse(w, msg, "Cloud integration saved successfully")
}

// UpdateCloudIntegration updates an existing cloud integration for a team.
// @Summary Update cloud integration
// @Description Updates a cloud integration for the given `teamId`. `integratorType` must be present in the request body.
// @Tags Integrations
// @Accept json
// @Produce json
// @Param teamId query string true "Team ID"
// @Param request body client.CreateIntegrationRequest true "Cloud integration request"
// @Success 200 {object} map[string]interface{}
// @Failure 400 {object} map[string]interface{} "Invalid request body or missing required parameters"
// @Failure 500 {object} map[string]interface{} "Internal server error"
// @Security ApiKeyAuth
// @Router /api/v1/integrations/cloud [put]
func (c *IntegratorController) UpdateCloudIntegration(w http.ResponseWriter, r *http.Request) {
	teamID := r.URL.Query().Get("teamId")
	if teamID == "" {
		utils.SendErrorResponse(w, http.StatusBadRequest, "teamId is required")
		return
	}

	var req client.CreateIntegrationRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		c.logger.LogError(err, "failed to decode cloud integration request", nil)
		utils.SendErrorResponse(w, http.StatusBadRequest, "invalid request body")
		return
	}

	if strings.TrimSpace(req.IntegratorType) == "" {
		utils.SendErrorResponse(w, http.StatusBadRequest, "integratorType is required in request body")
		return
	}

	msg, err := c.integratorService.UpdateCloudIntegration(r.Context(), teamID, &req)
	if err != nil {
		c.logger.LogError(err, "failed to update cloud integration", nil)
		utils.SendErrorResponse(w, http.StatusInternalServerError, err.Error())
		return
	}

	utils.SendSuccessResponse(w, msg, "Cloud integration updated successfully")
}

// ListIntegrations lists all active integrations
// @Summary List active integrations
// @Description Returns a list of all active integrations
// @Tags Integrations
// @Accept  json
// @Produce  json
// @Param   teamIds query string false "Comma-separated list of team IDs to filter by"
// @Param   integratorType query string false "Integrator type filter"
// @Success 200 {object} map[string]interface{}
// @Failure 400 {object} map[string]string "Invalid request"
// @Failure 500 {object} map[string]string "Internal server error"
// @Security ApiKeyAuth
// @Router /api/v1/integrations [get]
func (c *IntegratorController) ListIntegrations(w http.ResponseWriter, r *http.Request) {
	teamId := r.URL.Query().Get("teamIds")
	if teamId == "" {
		utils.SendErrorResponse(w, http.StatusBadRequest, "teamId is required")
		return
	}
	integratorType := r.URL.Query().Get("integratorType")
	integrators, err := c.integratorService.ListActiveIntegrations(r.Context(), integratorType, teamId)
	if err != nil {
		c.logger.LogError(err, "Failed to list integrations", nil)
		utils.SendErrorResponse(w, http.StatusInternalServerError, err.Error())
		return
	}
	utils.SendSuccessResponse(w, integrators, "Integrations listed successfully")
}

// ListIntegrationsForTeam lists all integrations for a team
// @Summary List integrations for a team
// @Description Returns a list of all integrations for a team
// @Tags Integrations
// @Accept  json
// @Produce  json
// @Param   teamId query string true "Team ID"
// @Success 200 {object} map[string]interface{}
// @Failure 400 {object} map[string]string "Invalid request"
// @Failure 500 {object} map[string]string "Internal server error"
// @Security ApiKeyAuth
// @Router /api/v1/integrations/team [get]
func (c *IntegratorController) ListIntegrationsForTeam(w http.ResponseWriter, r *http.Request) {
	teamId := r.URL.Query().Get("teamId")
	if teamId == "" {
		utils.SendErrorResponse(w, http.StatusBadRequest, "teamId is required")
		return
	}
	integrators, err := c.integratorService.ListIntegrationsForTeam(r.Context(), teamId)
	if err != nil {
		c.logger.LogError(err, "Failed to list integrations", nil)
		utils.SendErrorResponse(w, http.StatusInternalServerError, err.Error())
		return
	}
	utils.SendSuccessResponse(w, integrators, "Integrations listed successfully")
}

// DeleteIntegration deletes an integration.
// @Summary Delete integration
// @Description Deletes an integration by `integrationId` and `integrationName` within the provided `hubID`.
// @Tags Integrations
// @Accept json
// @Produce json
// @Param integrationId query string true "Integration ID"
// @Param integrationName query string true "Integration name"
// @Param hubID query string true "Hub ID"
// @Success 200 {object} map[string]interface{}
// @Failure 400 {object} map[string]interface{} "Missing required parameters"
// @Failure 500 {object} map[string]interface{} "Internal server error"
// @Security ApiKeyAuth
// @Router /api/v1/integrations/github/delete [delete]
func (c *IntegratorController) DeleteIntegration(w http.ResponseWriter, r *http.Request) {
	integrationId := r.URL.Query().Get("integrationId")
	if integrationId == "" {
		http.Error(w, "Integration ID is required", http.StatusBadRequest)
		return
	}

	hubID := r.URL.Query().Get("hubID")
	if hubID == "" {
		http.Error(w, "Hub ID is required", http.StatusBadRequest)
		return
	}

	integrationName := r.URL.Query().Get("integrationName")
	if integrationName == "" {
		http.Error(w, "Integration name is required", http.StatusBadRequest)
		return
	}

	err := c.integratorService.DeleteIntegration(r.Context(), integrationId, integrationName, hubID)
	if err != nil {
		c.logger.LogError(err, "Failed to delete integration", nil)
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}

	utils.SendSuccessResponseWithNoData(w, "Integration deleted successfully")
}